Moved AwsProxyExceptionHandler to proxy package. This should complete the refactoring of internal classes
